+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Unreleased]
+
+### Changed
+
+- PIN change uses `Notification` instead of `InfoMessage`. The notification is
+  less obtrusive, at the top of the screen.
+
 ## [2025.10-2] - 2025-10-24
 
 Yet another release with a big new feature, shortly after the last one 😅